ci: skip FFmpeg and pkgconf for ARM64 builds
FFmpeg and pkgconf prebuilt packages are not available for
Windows ARM64. Skip them in deps script.

+++ scripts/gha/deps_win32.sh
@@ -27,15 +27,17 @@ if [ "$GH_CPU_ARCH" = "i386" ]; then
 	rm yy-thunks.zip
 fi
 
-curl -L https://github.com/FWGS/potential-meme/releases/download/prebuilts/mingw-w64-x86_64-pkgconf-1.2.3.0-1-any.pkg.tar.zst -o pkgconf.tar.zst
-7z x pkgconf.tar.zst
-7z x pkgconf.tar
-rm pkgconf.tar*
-mv mingw64 pkgconf
+if [ "$GH_CPU_ARCH" != "arm64" ]; then
+	curl -L https://github.com/FWGS/potential-meme/releases/download/prebuilts/mingw-w64-x86_64-pkgconf-1.2.3.0-1-any.pkg.tar.zst -o pkgconf.tar.zst
+	7z x pkgconf.tar.zst
+	7z x pkgconf.tar
+	rm pkgconf.tar*
+	mv mingw64 pkgconf
 
-FFMPEG_ARCHIVE=$(get_ffmpeg_archive)
-curl -L "https://github.com/FWGS/FFmpeg-Builds/releases/download/latest/$FFMPEG_ARCHIVE.zip" -o ffmpeg.zip
-if [ -f ffmpeg.zip ]; then
-	unzip -x ffmpeg.zip
-	mv "$FFMPEG_ARCHIVE" ffmpeg
+	FFMPEG_ARCHIVE=$(get_ffmpeg_archive)
+	curl -L "https://github.com/FWGS/FFmpeg-Builds/releases/download/latest/$FFMPEG_ARCHIVE.zip" -o ffmpeg.zip
+	if [ -f ffmpeg.zip ]; then
+		unzip -x ffmpeg.zip
+		mv "$FFMPEG_ARCHIVE" ffmpeg
+	fi
 fi
